==> 02a_get_geo_data.ipynb <==
import geopandas as gpd
from shapely.geometry import Point, Polygon
from scipy import spatial
import numpy as np
import pandas as pd

#load índice bienestar territorial 2012
ibt = gpd.read_file('input/geo_data/ibt/ibt_geo.shp', crs={'init': 'epsg:5361'})
ibt = ibt.dropna(subset = ['nse']).reset_index(drop = True)
ibt.tail()

# calculate the centroid of each block
mcentroids = ibt['geometry'].centroid

# Create KDTree to consult the closest block per point
data = list(zip(mcentroids.x.ravel(), mcentroids.y.ravel()))
tree = spatial.KDTree(data = data, leafsize = 10)

# Import lat lon data
georef = pd.read_csv('input/deployment_data/posts_deployment.csv', encoding='utf-8')
georef.head()

# Isolate lat lon points for KDTree queries
pts = list(zip(georef['lon'].ravel(), georef['lat'].ravel()))
pts

%%time
# Calculate the 2 nearest blocks per lat lon point
kind = tree.query(pts, k = 2)[1]

%%time
# Calculate average NSE per block pair
georef.loc[:,('nse')] = list(map(lambda i: ibt.loc[kind[i],'nse'].mean(), range(len(pts))))

#output table, with lat lon data and an associated NSE
georef.to_csv('proc/02_posts_deployment_nse.csv', encoding = 'utf-8', index = False)
